pyproject.toml
[tool.poetry]
name = "docu-chat"
version = "0.1.0"
description = "AI-powered document question-answering system with RAG"
authors = ["Yasith Ranusha Silva <yasithranusha24@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
langchain = "^1.2.0"
langchain-community = "^0.4.1"
langchain-google-genai = "^4.1.2"
faiss-cpu = "^1.13.2"
pypdf = "^6.5.0"
python-dotenv = "^1.2.1"
langchain-text-splitters = "^1.1.0"
langchain-huggingface = "^1.2.0"
sentence-transformers = "^5.2.0"
python-multipart = "^0.0.21"
langchain-pinecone = "^0.2.13"
pinecone = ">=7.0.0,<8.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload"
start = "uvicorn app.main:app"